moved unit tests into their own folders
